Moved the SetUniform functions that where previously located in GL_ShaderProgam.h/.cpp to their own header and source file GL_Uniform.h/.cpp
