+++ OpenGL_Basic_Object_Transformations/Spinning_Cube/Debug/Spinning_Cube.log
@@ -1,7 +1,6 @@
 ﻿  GL_ShaderProgram.cpp
-  Math_Types.cpp
   Spinning_Cube.cpp
-C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Spinning_Cube.cpp(98,75): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
-C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Spinning_Cube.cpp(140,49): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
+C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Spinning_Cube.cpp(91,83): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
+C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Spinning_Cube.cpp(133,49): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
   Generating Code...
   Spinning_Cube.vcxproj -> C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Debug\Spinning_Cube.exe
