+++ OpenGL_Basic_Object_Transformations/Spinning_Cube/Debug/Spinning_Cube.tlog/Cl.items.tlog
@@ -2,6 +2,7 @@ C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningC
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\GL_ElementBuffer.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\GL_ElementBuffer.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\GL_ErrorHandeling.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\GL_ErrorHandeling.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\GL_ShaderProgram.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\GL_ShaderProgram.obj
+C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\GL_Uniform.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\GL_Uniform.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\GL_VertexArray.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\GL_VertexArray.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Gl_VertexBuffer.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\Gl_VertexBuffer.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Math_Types.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\Math_Types.obj
