+++ OpenGL_Basic_Object_Transformations/Spinning_Cube/Debug/Spinning_Cube.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_DRAW.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_ELEMENTBUFFER.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_ERRORHANDELING.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_SHADERPROGRAM.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_VERTEXARRAY.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_VERTEXBUFFER.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\MATH_TYPES.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\SPINNING_CUBE.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\WINDOW.OBJ
+^C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_DRAW.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_ELEMENTBUFFER.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_ERRORHANDELING.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_SHADERPROGRAM.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_UNIFORM.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_VERTEXARRAY.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\GL_VERTEXBUFFER.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\MATH_TYPES.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\SPINNING_CUBE.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_BASIC_OBJECT_TRANSFORMATIONS\SPINNING_CUBE\DEBUG\WINDOW.OBJ
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Basic_Object_Transformations\Spinning_Cube\Debug\Spinning_Cube.ilk
