+++ OpenGL_Basic_Object_Transformations/Spinning_Cube/GL_ShaderProgram.h
@@ -24,11 +24,4 @@ unsigned int CreateShaderProgram(const std::string& filepPath);						   // All s
 
 void UseShaderProgram(unsigned int shader_program); 
 
-GL_Uniform GetUniform(unsigned int shader_program, const std::string& u_Name);
-// Uniform set function overloading:
-void SetUniform1f(unsigned int shader_program, const std::string& u_Name, float data);
-void SetUniform1f(unsigned int shader_program, unsigned int u_Location,   float data);   // SHOULD BE FASTER TO SEARCH UNIFORMS BY HANDLE THAN BY NAME (GL DOCS) 
-void SetUniform4f(unsigned int shader_program, const std::string& u_Name, vec4f data); 
-void SetUniform4f(unsigned int shader_program, unsigned int u_Location,   vec4f data);   // SHOULD BE FASTER TO SEARCH UNIFORMS BY HANDLE THAN BY NAME (GL DOCS)
-
 #endif  // GL_SHADER_PROGRAM
