+++ OpenGL_Basic_Object_Transformations/Spinning_Cube/Spinning_Cube.vcxproj.filters
@@ -42,6 +42,9 @@
     <ClCompile Include="Math_Types.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="GL_Uniform.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GL_ErrorHandeling.h">
